+++ dev_scripts/qa.py
@@ -887,11 +887,6 @@ class QAUbuntu2204(QADebianBased):
     VERSION = "22.04"
 
 
-class QAUbuntu2304(QADebianBased):
-    DISTRO = "ubuntu"
-    VERSION = "23.04"
-
-
 class QAUbuntu2310(QADebianBased):
     DISTRO = "ubuntu"
     VERSION = "23.10"